+++ src/components/Header.stories.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import Header from './Header'
+
+export default {
+  title: 'Header',
+  component: Header,
+}
+
+const Template = () => <Header />
+
+export const Button = Template.bind({})
