--- src/components/HomeButton.stories.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import HomeButton from './HomeButton'
-
-export default {
-  title: 'HomeButton',
-  component: HomeButton,
-}
-
-const Template = () => <HomeButton />
-
-export const Button = Template.bind({})
