+++ src/hooks/useFilters.js
@@ -1,7 +1,14 @@
 import { useState } from 'react'
 
 export default function useFilters() {
-  const filters = ['Show all', 'Book', 'Audiobook', 'Podcast', 'Film', 'Series']
+  const filters = [
+    'Show all',
+    'Books',
+    'Audiobooks',
+    'Podcasts',
+    'Films',
+    'Series',
+  ]
   const [active, setActive] = useState(filters[0])
   function activeFilter(filter) {
     setActive(filter)
